GTMD有毒

+++ server/script/server-core.sh
@@ -4,15 +4,16 @@ source "${MINESH_SERVER_PATH}/script/errinfo.sh"
 source "${MINESH_SERVER_PATH}/script/utils.sh"
 source "${MINESH_SERVER_PATH}/script/config.sh"
 source "${MINESH_SERVER_PATH}/script/data-manager.sh"
+source "${MINESH_SERVER_PATH}/script/map.sh"
 
 tempPath="${MINESH_SVR_DATA_DIR}/.temp"
-requestQueue="%{tempPath}/queue"
+requestQueue="${tempPath}/queue"
 declare ncatPid
 
 # Established connections
 declare -A connections
-readonly CONNECTED=1
-readonly ACCEPTED=2
+readonly UNKNOWN=1
+readonly CONNECTED=2
 readonly GUEST=3
 readonly LOGEDIN=4
 
@@ -21,14 +22,56 @@ readonly LOGEDIN=4
 
 onServerStopped()
 {
+	if [[ -n $stopped ]]; then
+		return 0
+	fi
 	saveStateMap
 	mineshInfo "Server stopped."
 	rm -r $tempPath
+	if [[ -n $ncatPid ]]; then
+		kill $ncatPid
+	fi
+	stopped=1
+}
+
+stopServer()
+{
+	if [[ -z $stopped ]]; then
+		echo 'Stop' > ${requestQueue}
+		wait
+	fi
+}
+
+disconnect()
+{
+	echo "Disconnect" > "${tempPath}/$1"
+	$TPUT bold
+	$TPUT setaf 1
+	mineshLog "ID $1 disconnected."
+	$TPUT sgr0
 }
 
 serverLoop()
 {
+	local line
+	local request
+	while true; do
+		read request <&5
+		echoc 5 "Got request: $request"
+		if [[ $request = "Stop" ]]; then
+			for i in ${!connections[@]}; do
+				disconnect $i
+			done
+			onServerStopped
+			return 0
+		fi
 
+		line=$(cut -d ' ' -f 1 <<< "$request")
+		request=$(cut -d ' ' -f 2- <<< "$request")
+
+
+
+	done
 }
 
 # runServer [-p <port>]
@@ -45,6 +88,13 @@ runServer()
 
 	mkdir -p $tempPath
 
+
+	# Setup the log
+	setLogFile "${MINESH_SVR_DATA_DIR}/log"
+	$TPUT setaf 6
+	$TPUT bold
+	mineshLog "Server started"
+	$TPUT sgr0
 
 
 	# Load config file.
@@ -69,7 +119,7 @@ runServer()
 	# Use specified port if any.
 	if [[ $# -eq 2 ]]; then
 		if [[ $1 = "-p" ]]; then
-			if grep "^[0-9]+$" <<< $2; then
+			if grep -E -q "^[0-9]+$" <<< $2; then
 				port=$2
 			else
 				mineshErr "Invalid port. Default port will be used."
@@ -77,11 +127,11 @@ runServer()
 		fi
 	fi
 
-	if ! loadMap; then
-		mineshErr "Load map file failed."
-		onServerStopped
-		return 1
-	fi
+	# if ! loadMap "${MINESH_SVR_DATA_DIR}/map"; then
+	# 	mineshErr "Load map file failed."
+	# 	onServerStopped
+	# 	return 1
+	# fi
 
 	ncat --version
 	if [[ $? -ne 0 ]]; then
@@ -100,19 +150,26 @@ runServer()
 	ncat -vlk --sh-exec "$cmd" $port &
 	ncatPid=$!
 
+
 	serverLoop &
 
 	local loopPid=$!
 
-	trap "echo 'stop' > ${requestQueue}; wait; exit 1" 0 1 2 3 15
+	trap "stopServer; exit 1" 1 2 3
+	trap "stopServer; exit 0" 15
 	
 	$TPUT bold
 	mineshInfo "Server started. Press Ctrl+D to stop."
-	while read -s -n 1; do
-		:
+	$TPUT sgr0
+
+	local CtrlD=$(echo -e '\x04')
+
+	while read -s -n 1 ch; do
+		if [[ $ch = $CtrlD ]]; then
+			break
+		fi
 	done
 
-	echo 'stop' > ${requestQueue}
-	wait
+	stopServer
 	return 0
 }
